requirements.txt
bidict==0.23.1
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
simple-websocket==1.0.0
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzdata==2024.1
Werkzeug==3.0.1
wsproto==1.2.0
